+++ src/com/intellij/idea/plugin/hybris/startup/HybrisProjectStructureStartupActivity.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -101,7 +101,6 @@ class HybrisProjectStructureStartupActivity : ProjectActivity {
         resetSpringGeneralSettings(project)
         fixBackOfficeJRebelSupport(project)
 
-        CommonIdeaService.instance.fixRemoteConnectionSettings(project)
         ConsolePersistenceService.getInstance(project).loadPersistedQueries()
     }